Move Markdown component files to components folder
